requirements.txt
Django~=3.1.0
mysqlclient
pypugjs
gunicorn
django-crispy-forms
